+++ src/app/posts/new/_components/input-image-with-preview/index.ts
@@ -1,0 +1,1 @@
+export * from './input-image-with-preview';
