+++ src/app/posts/new/_components/input-image-with-preview/input-image-with-preview.stories.tsx
@@ -1,0 +1,13 @@
+import { InputImageWithPreview } from './input-image-with-preview';
+
+import type { Meta, StoryObj } from '@storybook/react';
+
+export default {
+  component: InputImageWithPreview,
+} as Meta<typeof InputImageWithPreview>;
+
+type Story = StoryObj<typeof InputImageWithPreview>;
+
+export const Default: Story = {
+  args: {},
+};
