+++ src/app/posts/new/_components/input-image-with-preview/input-image-with-preview.tsx
@@ -1,0 +1,44 @@
+'use client'
+import clsx from 'clsx';
+
+import * as styles from './input-image-with-preview.css';
+
+import { useState, type FC, type InputHTMLAttributes } from 'react';
+
+import Image from 'next/image';
+
+const ACCEPTABLE_TYPES = 'image/png, image/jpeg, image/jpg, image/gif';
+
+export type InputImageWithPreviewProps = {
+  className?: string | undefined;
+  name: InputHTMLAttributes<HTMLInputElement>['name'];
+};
+
+
+const imageUrl = (image: File) => {
+  if (image === undefined) return '';
+  return window.URL.createObjectURL(image);
+};
+
+export const InputImageWithPreview: FC<InputImageWithPreviewProps> = ({
+  className,
+  name,
+}) => {
+  const [image, setImage] = useState<File>();
+
+  const onFileInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (!e.target.files) return;
+
+    const fileObject = e.target.files[0];
+    if (fileObject === undefined) return;
+
+    setImage(fileObject);
+  };
+
+  return (
+    <div className={clsx(className, styles.wrapper)}>
+      <input type="file" name={name} accept={ACCEPTABLE_TYPES} onChange={onFileInputChange} />
+      {image && <Image src={imageUrl(image)} alt="Preview" width='320' height='300' />}
+    </div>
+  );
+};
