--- src/app/posts/new/_components/preview-image/index.ts
@@ -1,1 +1,0 @@
-export * from './preview-image';
